refactor: components to folders
