+++ src/components/Card/Card.tsx
@@ -1,4 +1,4 @@
-import type { CardProps } from '../types';
+import type { CardProps } from '../../types';
 import './card.css';
 
 const Card = ({ item }: CardProps) => (
