+++ src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,5 +1,5 @@
 import { Component, ErrorInfo } from 'react';
-import type { ErrorBoundaryProps } from '../types';
+import type { ErrorBoundaryProps } from '../../types';
 
 class ErrorBoundary extends Component<ErrorBoundaryProps> {
   state = { hasError: false };
